backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-decouple
pymongo